some progress making AnimatronLib be able to render individual frames many todos, but actually not that far off from being editor useful.
