+++ config.h
@@ -1,8 +1,12 @@
+
+// only for reporting to the user
 static const int c_programVersionMajor = 0;
 static const int c_programVersionMinor = 5;
 
+// the current version of the document
 static const int c_documentVersionMajor = 1;
 static const int c_documentVersionMinor = 0;
 
+// the current version of the config
 static const int c_configVersionMajor = 1;
 static const int c_configVersionMinor = 0;
